fix(suite-native): account rename handler
# Conflicts:
#	suite-native/labeling/src/components/AddressLabelEditable.tsx
#	suite-native/labeling/src/components/TransactionOutputLabelEditable.tsx
#	suite-native/module-accounts-management/src/components/AccountRenameForm.tsx
