+++ suite-native/labeling/src/components/AddressLabelEditable.tsx
@@ -55,9 +55,8 @@ export const AddressLabelEditable = ({
                     showToast({ variant: 'error', icon: 'warning', message: type });
 
                     return;
-
                 case 'WriteModeRequiredForAllocation':
-                    // expected case
+                    // Do nothing, this is expected control flow error when we want allocate on-demand.
                     return;
                 default:
                     return exhaustive(type);
