+++ suite-native/trading-state/package.json
@@ -27,6 +27,7 @@
         "@suite-native/accounts": "workspace:*",
         "@suite-native/feature-flags": "workspace:*",
         "@suite-native/labeling": "workspace:*",
+        "@suite-native/settings": "workspace:*",
         "@suite-native/tokens": "workspace:*",
         "@suite-native/trading-atoms": "workspace:*",
         "@suite-native/trading-consts": "workspace:*",
