+++ suite-native/trading-state/tsconfig.json
@@ -36,6 +36,7 @@
         { "path": "../accounts" },
         { "path": "../feature-flags" },
         { "path": "../labeling" },
+        { "path": "../settings" },
         { "path": "../tokens" },
         { "path": "../trading-atoms" },
         { "path": "../trading-consts" },
